CNN model & and augmentation

+++ duplicates.py
@@ -5,16 +5,12 @@ import imagehash
 from tqdm import tqdm
 from multiprocessing import Pool, Manager
 import multiprocessing as mp
+import pandas as pd
 
-output_folder = 'output'
+output_folder = r"output"
 image_extensions = ['*.jpeg', '*.jpg', '*.png']
-
-
-image_files = []
-for ext in image_extensions:
-    image_files.extend(glob.glob(os.path.join(output_folder, ext)))
-
-print(f"Found {len(image_files)} images in the folder {output_folder}")
+csv_path = r'database\trainLabels.csv'
+updated_csv_path = r'database\trainLabels_updated.csv'
 
 def compute_hash(image_path):
     try:
@@ -45,15 +41,36 @@ def parallel_hash_computation(image_files):
     return hash_dict
 
 def remove_duplicates():
+    print("\nDuplicate removal stage")
+    print("------------------------")
+    image_files = []
+    for ext in image_extensions:
+        files = glob.glob(os.path.join(output_folder, ext))
+        image_files.extend(files)
+    print(f"Found {len(image_files)} images in {output_folder}")
+
+    if not image_files:
+        print("No images found in output folder!")
+        return
+
     hash_dict = parallel_hash_computation(image_files)
 
     duplicates_count = 0
+    kept_files = []
     for img_hash, file_list in hash_dict.items():
+        kept_files.append(file_list[0]) # Keep the first file
         if len(file_list) > 1:
             duplicates_count += len(file_list) - 1
             for file_to_remove in file_list[1:]:
-                print(f"Removing duplicate: {file_to_remove}")
+                # print(f"Removing duplicate: {file_to_remove}")
                 os.remove(file_to_remove)
 
     print(f"Found and removed {duplicates_count} duplicates.")
     print(f"{len(image_files) - duplicates_count} unique images remain.")
+
+    # Aktualizacja pliku CSV
+    df = pd.read_csv(csv_path)
+    kept_filenames = [os.path.splitext(os.path.basename(f))[0] for f in kept_files]
+    updated_df = df[df['image'].isin(kept_filenames)]
+    updated_df.to_csv(updated_csv_path, index=False)
+    print(f"Updated CSV saved to {updated_csv_path} with {len(updated_df)} entries.")
